% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_margin.R
\name{add.margin}
\alias{add.margin}
\title{Sample Size Calculation for Non-Inferiority Trial for Additive Mean Hypothesis}
\usage{
add.margin(delta, varC, varT = NULL, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{delta}{a numerical value of the additive value in the hypothesis ($\Delta > 0$)}

\item{varC}{a numerical value of the variance of the control group}

\item{varT}{a numerical value of the variance of the treatment group, it is null if it is same as varC}

\item{alpha}{Two-sided Type-I-error}

\item{beta}{Type-II-error}
}
\value{
a list of sample sizes for control, treatment, and randomization ratio for treatment to control and efficiency compared 1:1 randomization ratio
}
\description{
This function computes the required sample size for non-inferiority trial with additive margin.
It also computes the sample size required for both treatment and control group.
It also computes the efficiency compared to equal treatment allocation.
}
\author{
Thevaa Chandereng, Rick Chappell
}
